+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 * Add Array#closest(goal) method to find the closest value to the given goal
 * Added String#before(needle, first) and String#after(needle, first)
 * Added String#beforeLast(needle) and String#afterLast(needle)
+* Added Object.isObject(obj)
+* Added Object.isPlainObject(obj)
+* Added Object.isPrimitiveObject(obj)
 
 ## 0.1.1 (2014-06-23)